Animations implementation continued...

+++ Car(d)Battle/Cards.cpp
@@ -113,7 +113,7 @@ void Cards::print()
 std::string Cards::get_category_name(int num)
 {
 	num++;
-	if (num >= field_names.size() || num < 0)
+	if (static_cast<unsigned int>(num) >= field_names.size() || num < 0)
 		return std::string();
 	return field_names[num];
 }
